Made barely functioning client-side encryption, needs heavy polishing
Client-side encryption now works (SubtleCrypto was used). However, there is an issue with the message processing where the output is sometimes altered. For example, if the message "hello, world" is sent, the output may be "helloworl". Similarly, instead of receiving "1234567890", "123456789w==" is received. Additionally, React's setState is not functioning properly, and the code needs to be refined.